requirements.txt
amqp==5.3.1
anyio==4.9.0
asgiref==3.8.1
attrs==25.3.0
billiard==4.2.1
black==25.1.0
celery==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.0
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.5
django==5.2.1
django-celery-beat==2.8.1
django-celery-results==2.6.0
django-debug-toolbar==5.2.0
django-filter==25.1
django-redis==5.4.0
django-stubs==5.2.0
django-stubs-ext==5.2.0
django-timezone-field==7.1
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
drf-spectacular==0.28.0
environs==14.1.1
flower==2.0.1
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
humanize==4.12.3
idna==3.10
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
kombu==5.5.3
markdown==3.8
marshmallow==4.0.0
mypy==1.15.0
mypy-extensions==1.1.0
packaging==25.0
pathspec==0.12.1
pillow==11.2.1
platformdirs==4.3.8
prometheus-client==0.22.0
prompt-toolkit==3.0.51
psycopg2-binary==2.9.10
pyjwt==2.9.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-telegram-bot==22.1
pytz==2025.2
pyyaml==6.0.2
redis==6.1.0
referencing==0.36.2
requests==2.32.3
rpds-py==0.25.0
ruff==0.11.10
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
stripe==12.1.0
tornado==6.5
types-pyyaml==6.0.12.20250516
typing-extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
